Initial commit of basic si5324 driver code
Basic structure has been added, and some functions fleshed out.

Individual fields within registers can be set, and building on this,
several registers can be set at a time when a file generated by the
DSPLLsim software is used. However, this has not been tested.

ICAL is required when certain registers are modified, so this will
be automatically performed if needed (unless otherwise specified).