Added functions for alarm checking and clock selection
get_alarm_states will now report the status of alarm states (interrupts) and flags
when called externally, and grabs all values at the same time to reduce the number
of reads from reigsters that are shared. This has also involved the addition of the
Alarms class, of which an instance is returned by the function.

There is also a new function for selecting the clock that drives the PLL, which
will automatically set freerun mode on and select clock 2 when the oscillator is
chosen. There is also a complementary set_autoselect_mode function that is used
to switch between automatic (priority) and manual selection. If the manual mode
is not selected when a clock is specified with the first function, a warning
will be shown.