Added timeout for iCAL LOL check
When iCAL is run, it will cause the PLL LOL (Loss Of Lock) flag to remain high
for the duration. When calibration is complete, the PLL lock should be found
again, and the LOL INT will go low again.

Until now, the code for triggering the iCAL waited indefinitely for the LOL INT
to go low again. However, there are circumstances where this could be never,
as I have found when setting the input clock to one that has no signal. To
avoid being stuck in the calibrate() function indefinitely, it now includes
an (adjustable) timeout_ms value that will throw a warning if exceeded. This
warning will alert the user and print out alarms so that possible LOS (Loss
of Signal) alarms can be seen for the chosen channel. They are also prompted
to run the iCAL again once the signal is found, since running without iCAL
can result in an unstable output.