Add tests for checking si5324 device driver
A test script has now been added to test the functionality of the si5324 driver
by mocking the I2C layer. Some I2C functions are overwritten to write to a live
register map, so that changes to these registers can be checked, and initial
starting configurations for the si5324 can be mocked.

Tests include:
    * Device I2C address selection from pins
    * Calibration routine
    * Calibration requirement tracking
    * Register field (_Field) read and write
    * Input clock selection for 1, 2 and X (inc freerun)
    * Active clock reading
    * Register map file import and export
    * Alarm state reading (INTs and FLGs)