Updates to pass python2 build
The open() function from builtins was being overridden in the
test script, but this module has had a name change between version.
This has been dealt with by version in the test script.

Although the Python3 built did not pick it up, the python2 test needs
static functions to be declared, so this has been done for the si5324
pin address function within class SI5324.