Created channel inner class for device channels 0-3 and all
Originally, channel number was going to be passed to register access functions
and be ignored (or a default value assumed) if it was not needed.

This has now been updated to have a separate class for channel-related functions,
with the ds110 instance having 5 instances as attributes (4 channels including an
all channel for writing to all channels). This will make the divide between control
of device-specific and channel-specific settings more clear.

Some functionality has been added to test this structure out, including reset
functions for both the channel registers and shared control registers.