Add bulk of CDR settings and config
Ideally the logic included in these functions should be checked by someone with
experience using the device, as the datasheet leaves many details unclear or omits
explanation of fields entirely.

More work is also needed on the manual rate and subrate function, which either needs
to calculate the expected divider ratios automatically selected, or in the case where
automatic dividers may be changed allow a manual PPM value to be specified, assuming
the user will know the expected divider ratio that will be used out of the specified
selection.