Update CDR autoconfig function to calc settings to achieve data rates
The manual (non-standards-based) autoconfig function now determines correct settings
required to achieve the desired data rates either by being supplied a VCO frequency
and calculating the needed divider ratios (if possible) or vice versa.

Some of the direct access field writing parts of this function have also been moved
out to individual functions so that that can be called if (for example) values for
the PPM count registers are already known.