Basics of a unified FireFly driver laid out as a concept
The simplest form of an auto-detecting unified FireFly driver has been added. This
driver will read the upper byte 128 to determine if QSFP or CXP is being used, set
up this interface with a chosen address if given (using a supplied selectL line if
multiple devices are on the same bus), and allow the access of fields to be made
abstract from the differing memory maps the two standards have.

This means the driver can focus on the operations themselves. Examples given are
reading temperatures, getting vendor information, and disabling channels. On init
the driver will disable all channels to protect the device from immediately
overheating.