Rewrote interface identification and part number deconstruction
Previously the interface identification assumed that the SFF identifiers that should be
present in QSFP+ and CXP would be offered by the FireFlies, but the CXP-based device
does not give a valid result. The decision is therefore now made by hunting down the
vendor OUI supported in the CXP version, which holds the part number in the QSFP version.

The part number is also now deconstructed and analysed, with some useful information used
by the driver (e.g. number of channels). If the value extracted does not comply with the
expectations, either an exception will be raised or a warning given depending on the impact
on driver usage.