Added basics of GPIO handling with gpiod bindings
As it stands, the code allows the definition of system GPIO buses with a given offset
unique to the architecture. Examples have already been included for ZynqMP and Zynq,
currently with a width of 12. However, this could be chanegd.

Custom buses can be defined in the same way without issue, as long as the pins have
not yet been requested.

Pins can be requested and then written/read individually by bus index, and can also
be requested in bulk and then written to / read from as an array.

There is currently no support for event detection, but this will be added in the
future by using asyncio for concurrency.