Support for asynchronous events added, untested on HW
Asynchronous events have now been added using concurrent.futures to create tasks in a
ThreadPool for each monitor. These monitors are self-terminating when requested through
the remove_pin_event_callback() function. The user is able to supply their own callback.

Pins are also now checked for availability before requests are tried using the is_used()
function. This requires the pin information to be in sync with the kernel, which requires
the Line.update() function to be available (which it is not in earlier versions of gpiod).
This is checked at import. The version of libgpiod has been updated in the CFI DAQ
PetaLinux control project to meet this requirement.