Added synchronous pin event handling and initial test script fixes
Synchronous functions for requesting event monitoring with gpiod have been added so that the
controlling code can simply poll the wait_pin_event function with a timeout if required. This
means it can be used where concurrent.futures is not available, or if the code model is not
asynchronous.

The initial tests have also been added, and several corrections made based on errors found.