Attempt to fix errors with python2 CI
Most issues with Travis on Python2 seem to be the lack of presence of ModuleNotFoundError,
which is now set to ImportError for python2.

Also, there is no concurrent.futures module for python2. There is, however, futures. Therefore the
driver code now imports whichever is available as futures if it is there.