Add basics of SI534x driver and related classes
There are classes for fields, as well as derived classes for channel-mapped registers (which are
offset using a channel number depending on the number of channels for the current device).
Similarly, multisynth-mapped register fields are offset by a certain multisynth number, of which
there are different total numbers in different devices.