Update channel-mapped registers to support specified map positions
It turns out that the devices with less channels have their numbered channels dispersed
through the map for the 10 channels present with the SI5345, though otherwise the maps
are compatible. For example, the SI5344 uses the same channel-mapped fields as the
SI5345, but uses the positions 2, 3, 6 and 7 for its channels 0, 1, 2 and 3.

This means that when writing a channel for a driver, it is supplied the logical channel
number, which is converted to a channel position, which then determines the offset from
the SI5345 channel 0 register that should be used for the given operation.