Moved export register definition to an iterator generator class
The register definitions for the register value export process are now generated from the original list
of ranges (and exclusions) using a generator class that provides a resulting simple tuple iterable.

This may seem over-complicated, but it simlifies the modification of the allowed register extraction
map, while making the export function simpler.