Updated function names to suit scope
The functions written for I2C read/write were actually mostly generic, and so have been renamed
to serve both the SPI and I2C interfaces. The calls to each individual interface type are now
wrapped and simplified so that they only require a register and value for write. The page
handling is done at the calling layer.