Added support for driving outputs individually
This commit should provide enough functionality for MERCURY. Outputs can be set one at a time,
and the write-thru capability of the device has been left out for now. Additionally, the nLDAC
pin is not driven to latch values from the input to DAC registers, and SPI commands for
latching are used instead.