Post-testing fixes
Previously, the power-up for the output being set had not been implemented, but it turns out
that channels start in shutdown-3 on POR. Therefore, new functions for setting individual
output power level have been added, and unless specifically told not to, the set_output()
function will also power up the specified output.

Additionally, the writebytes2 function used in the SPI device driver (which comes from spidev)
is not currently available on the version of spidev included in the PetaLinux build by default
There will need to be further investigation into adding this in the future, but for now, the
call to write_16() has been replaced with an equaivalent transfer() call with a discarded
return value.