Added read_field test and fixed read_field issues
Issues with read_field were to do with the number of output bytes being derived incorrectly. However, it only failed if the number of bytes output was differernt to the number of registers being read (e.g. reading 4 bits accross a byte boundary), so had not been used so far.