Updated I2C transactions to use i2c_rdwr due to nonstandard structure
This device is actually quite difficult to use due to it's no-standard
control interface via I2C, which does not use typical register-based
SMBus commands. It also requires that certain write-read transactions
are executed concurrently without any other traffic on the bus (no
matter the address), which is difficult to ensure.