Fix test in other file by using MagicMock rather than Mock
Oddly, using Mock() for smbus actually broke the test in the ad5272 file
that was executed after this test. Changing the filename to make this
test occur last also fixed this issue, so the ad5259 test was actually
interfering with the other device's test file.

Changing smbus to MagicMock() appears to sort the issue. It seems
tests are not as segregated as I'd imagined.