FireFly Addressing Overhaul
FireFly addressing can now properly account for different initial states
of FireFlies- since the settings of different addresses are actually
non- volatile. A user can specify a base_address for the address that
the device is currently known to be on to initiate communication, and
also an optional chosen_base_address, which will reprogram the device to
another address (exact limits of this depend on the interface type).