Add support for FireFly parts
Previously, certain information was always decoded from the device part
number (most critically number of channels, and device direction). This
works well for most parts. However, custom parts have their part number
replaced with a part starting 'OTP...', which does NOT follow the part
number pattern set out in the ordering information. These fields also
cannot be decoded from an OTP part number.

Therefore, if using a custom part, it is now possible (required) to
supply the direction and number of channels manually. If these are not
given, a sensible error is now shown rather than a cryptic failure to
decode the part number. It will be up to the use to know these details
for the device in use.

For no-custom parts, everything will work as normal, and no additional
parameters are required.