Fix/firefly address fixes (#26)
* Allow fireflies to be created with user specificed I2C bus

* FireFly Addressing Overhaul

FireFly addressing can now properly account for different initial states
of FireFlies- since the settings of different addresses are actually
non- volatile. A user can specify a base_address for the address that
the device is currently known to be on to initiate communication, and
also an optional chosen_base_address, which will reprogram the device to
another address (exact limits of this depend on the interface type).

* Minor log changes

* Add default direction for reading duplex device temperature without throwing an error

* Add support for FireFly parts

Previously, certain information was always decoded from the device part
number (most critically number of channels, and device direction). This
works well for most parts. However, custom parts have their part number
replaced with a part starting 'OTP...', which does NOT follow the part
number pattern set out in the ordering information. These fields also
cannot be decoded from an OTP part number.

Therefore, if using a custom part, it is now possible (required) to
supply the direction and number of channels manually. If these are not
given, a sensible error is now shown rather than a cryptic failure to
decode the part number. It will be up to the use to know these details
for the device in use.

For no-custom parts, everything will work as normal, and no additional
parameters are required.